backend/requirements.txt
flask
google-cloud-storage
python-dotenv
pyspark
numpy